+++ tests/Queries/CachePeopleEmailQuery.php
@@ -4,10 +4,13 @@ namespace Sfneal\Datum\Tests\Queries;
 
 use Illuminate\Database\Eloquent\Model;
 use Sfneal\Datum\Tests\Models\People;
+use Sfneal\Datum\Tests\Queries\Traits\PeopleBuilder;
 use Sfneal\Queries\CacheModelQuery;
 
 class CachePeopleEmailQuery extends CacheModelQuery
 {
+    use PeopleBuilder;
+
     /**
      * Target Model.
      *