+++ tests/Queries/PeopleQueryWithFilters.php
@@ -2,15 +2,16 @@
 
 namespace Sfneal\Datum\Tests\Queries;
 
-use Illuminate\Database\Eloquent\Builder;
 use Sfneal\Datum\Tests\Filters\CityFilter;
 use Sfneal\Datum\Tests\Filters\NameFirstFilterDynamic;
 use Sfneal\Datum\Tests\Filters\NameLastFilter;
-use Sfneal\Datum\Tests\Models\People;
-use Sfneal\Queries\AbstractQueryWithFilters;
+use Sfneal\Datum\Tests\Queries\Traits\PeopleBuilder;
+use Sfneal\Queries\AbstractFilterableQuery;
 
-class PeopleQueryWithFilters extends AbstractQueryWithFilters
+class PeopleQueryWithFilters extends AbstractFilterableQuery
 {
+    use PeopleBuilder;
+
     /**
      * Retrieve an array of model attribute keys & corresponding Filter class values.
      *
@@ -23,22 +24,5 @@ class PeopleQueryWithFilters extends AbstractQueryWithFilters
             'name_last' => NameLastFilter::class,
             'name_first' => NameFirstFilterDynamic::class,
         ];
-    }
-
-    /**
-     * Execute a DB query using filter parameters.
-     *
-     * @return Builder
-     */
-    public function execute(): Builder
-    {
-        // Initialize query
-        $query = People::query();
-
-        // Apply filters
-        $query = $this->applyFiltersToQuery($query);
-
-        // Return query
-        return $query;
     }
 }